+++ lib/logster/json_formatter.ex
@@ -1,7 +1,0 @@
-defmodule Logster.JSONFormatter do
-  def format(params) do
-    params
-    |> Enum.into(%{})
-    |> Jason.encode_to_iodata!()
-  end
-end